requirements.txt
annotated-doc==0.0.3
annotated-types==0.7.0
anyio==4.11.0
click==8.3.0
colorama==0.4.6
fastapi==0.121.0
h11==0.16.0
idna==3.11
pydantic==2.12.3
pydantic-core==2.41.4
sniffio==1.3.1
starlette==0.49.3
typing-extensions==4.15.0
typing-inspection==0.4.2
uvicorn==0.38.0
pandas==2.2.3
gspread==5.12.3
unidecode==1.3.0
oauth2client==4.1.3
google-api-python-client==2.147.0
google-auth==2.35.0
google-auth-oauthlib==1.2.1
google-auth-httplib2==0.2.0
httplib2==0.22.0


python-dotenv==1.0.1
openpyxl==3.1.5